add missing idf targets
