+++ package.json
@@ -493,17 +493,24 @@
             "default": "esp32",
             "enum": [
               "esp32",
-              "esp32s2",
-              "esp32s3",
+              "esp32c2",
               "esp32c3",
               "esp32c6",
+              "esp32h2",
+              "esp32p4",
+              "esp32s2",
+              "esp32s3",
               "custom"
             ],
             "enumDescriptions": [
               "esp32",
-              "esp32 s2",
-              "esp32 s3",
-              "esp32 c3",
+              "esp32 C2",
+              "esp32 C3",
+              "esp32 C6",
+              "esp32 H2",
+              "esp32 P4",
+              "esp32 S2",
+              "esp32 S3",
               "custom"
             ],
             "description": "%param.adapterTargetName%",
@@ -1266,7 +1273,7 @@
       },
       {
         "command": "espIdf.qemuCommand",
-        "title": "QEMU Manager",
+        "title": "%espIdf.launchQemu.title%",
         "category": "ESP-IDF"
       },
       {
